fixup(weather): move modules into separate files
